+++ lib/builder/builder_functions.dart
@@ -2,5 +2,5 @@
 import 'package:flutter/widgets.dart';
 
 typedef Widget DataBuilder<T>(BuildContext context, T data);
-typedef Widget ErrorBuilder(BuildContext context, Object error);
+typedef Widget ErrorBuilder(BuildContext context, Object? error);