+++ CHANGELOG.md
@@ -1,4 +1,6 @@
-## [0.0.2+] - Fix future builder
+## [0.1.0-nullsafety] - Upgrade to Null-Safety
+
+## [0.0.2+1] - Fix future builder
 
 ## [0.0.2] - Add more options to all widgets